requirements.txt
absl-py==0.8.1
astor==0.8.0
astroid==2.2.5
attrs==19.3.0
cachetools==3.1.1
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Click==7.0
colorama==0.4.1
comet-git-pure==0.19.13
comet-ml==2.0.17
configobj==5.0.6
dominate==2.4.0
einops==0.1.0
everett==1.0.2
Flask==1.1.1
Flask-Bootstrap4==4.0.2
Flask-Script==2.0.6
Flask-WTF==0.14.2
gast==0.2.2
google-auth==1.7.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.25.0
gunicorn==20.0.0
h5py==2.10.0
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.0.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
lazy-object-proxy==1.4.2
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
memcnn==1.0.0
netifaces==0.10.9
numpy==1.17.3
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
opt-einsum==3.1.0
pep8==1.7.1
Pillow==6.2.1
protobuf==3.10.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycparser==2.19
pylint==2.3.1
pyrsistent==0.15.5
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
SimpleITK==1.2.3
six==1.12.0
tensorboard==2.0.1
tensorboardX==1.4
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
torch==1.3.1
torchvision==0.4.1
tqdm==4.38.0
typed-ast==1.4.0
urllib3==1.25.6
visitor==0.1.3
websocket-client==0.56.0
Werkzeug==0.16.0
wincertstore==0.2
wrapt==1.11.2
WTForms==2.2.1
wurlitzer==2.0.0
